Fix play overlay cosmetic bug and youtube playback in details page.
