+++ app/src/main/java/com/codepath/flickster/activities/MovieDetailActivity.java
@@ -11,40 +11,23 @@ import android.widget.TextView;
 
 import com.codepath.flickster.R;
 import com.codepath.flickster.adapters.Movie;
+import com.codepath.flickster.util.YouTubeUtil;
+import com.google.android.youtube.player.YouTubeBaseActivity;
+import com.google.android.youtube.player.YouTubePlayerView;
 import com.squareup.picasso.Picasso;
 
-public class MovieDetailActivity extends AppCompatActivity {
-
-    private class ImageOnClickListener implements AdapterView.OnClickListener {
-        private Movie movie;
-
-        private ImageOnClickListener(Movie movie) {
-            this.movie = movie;
-        }
-
-        @Override
-        public void onClick(View item) {
-            Intent intent = new Intent(getApplicationContext(), YouTubeActivity.class);
-            intent.putExtra("movie", movie);
-            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            getApplicationContext().startActivity(intent);
-        }
-    }
+public class MovieDetailActivity extends YouTubeBaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         Movie movie = (Movie) getIntent().getSerializableExtra("movie");
-        String imageBaseURL = getIntent().getStringExtra("imageBaseURL");
-        final String POSTER_IMAGE_SIZE = getIntent().getStringExtra("POSTER_IMAGE_SIZE");
-        final String BACK_DROP_IMAGE_SIZE = getIntent().getStringExtra("BACK_DROP_IMAGE_SIZE");
 
         setContentView(R.layout.activity_movie_detail);
 
         TextView title = (TextView) findViewById(R.id.tvMovieDetailTitle);
         TextView overview = (TextView) findViewById(R.id.tvMovieDetailOverview);
         TextView releaseDate = (TextView) findViewById(R.id.tvMovieDetailReleaseDateVal);
-        ImageView imageView = (ImageView) findViewById(R.id.ivMovieDetailImage);
         RatingBar ratingBar = (RatingBar) findViewById(R.id.rbMovieDetail);
 
         title.setText(movie.getTitle());
@@ -52,14 +35,7 @@ public class MovieDetailActivity extends AppCompatActivity {
         releaseDate.setText(movie.getReleaseDate());
         ratingBar.setRating((float) movie.getVoteAverage());
 
-        imageView.setOnClickListener(new ImageOnClickListener(movie));
-
-        //Use poster image if movie does not have backdrop image.
-        String imageUri = imageBaseURL +
-                (movie.getBackdropImagePath() != null && ! "null".equals(movie.getBackdropImagePath()) ?
-                        BACK_DROP_IMAGE_SIZE + "/" + movie.getBackdropImagePath() :
-                        POSTER_IMAGE_SIZE + "/" +movie.getPosterImagePath());
-        Picasso.with(this).load(imageUri)
-                .placeholder(R.drawable.ic_movie_placeholder).into(imageView);
+        YouTubePlayerView youTubePlayerView = (YouTubePlayerView) findViewById(R.id.movieDetailPlayer);
+        YouTubeUtil.init(movie, youTubePlayerView, true);
     }
 }
