+++ app/src/main/res/layout/item_movie_popular.xml
@@ -5,16 +5,14 @@
     <ImageView
         android:id="@+id/ivMovieImage"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"/>
+        android:layout_height="wrap_content"/>
 
     <ImageView
         android:id="@+id/playIconOverlay"
         android:src="@drawable/play_overlay"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"/>
+        android:layout_width="@dimen/play_overlay_size"
+        android:layout_height="@dimen/play_overlay_size"
+        android:layout_centerVertical="true"
+        android:layout_centerHorizontal="true"/>
 
 </RelativeLayout>
